make html not * the default selector
Currently if you leave off the selector argument the cli breaks. It
would be nice to have a sane default, like print the whole html
document. There are some places where Options are converted into
Strings, implying that * is supposed to be the default selector. This
change realises that implied intention, but decides html, which prints
the whole document, instead of *, which prints every tag recursively, is
a better default.